finish build rename
